+++ packages/node-core/src/indexer/mmrMigrate.service.ts
@@ -5,10 +5,10 @@ import {existsSync} from 'fs';
 import {DEFAULT_WORD_SIZE} from '@subql/common';
 import {FileBasedDb} from '@subql/x-merkle-mountain-range';
 import {Logging, QueryTypes, Sequelize} from '@subql/x-sequelize';
-import {NodeConfig} from '../configure/NodeConfig';
+import {NodeConfig} from '../configure';
+import {PgBasedMMRDB} from '../indexer/entities/Mmr.entitiy';
 import {getLogger} from '../logger';
 import {getExistingProjectSchema} from '../utils/project';
-import {PgBasedMMRDB} from './entities/Mmr.entitiy';
 
 const logger = getLogger('mmr-migrate');