+++ packages/node-core/src/indexer/project.service.ts
@@ -17,7 +17,7 @@ import {MetadataKeys} from './entities';
 import {MmrService} from './mmr.service';
 import {PoiService} from './poi/poi.service';
 import {StoreService} from './store.service';
-import {CacheMetadataModel} from './storeCache/cacheMetadata';
+import {CacheMetadataModel} from './storeCache';
 import {IProjectNetworkConfig, IProjectService, ISubqueryProject} from './types';
 import {IUnfinalizedBlocksService} from './unfinalizedBlocks.service';