+++ packages/utils/src/query/types.ts
@@ -17,16 +17,6 @@ export type MetaData = {
   indexerNodeVersion: string;
   queryNodeVersion: string;
   startHeight?: number;
-  rowCountEstimate: [TableEstimate];
+  rowCountEstimate: TableEstimate[];
   deployments: Record<number, string>;
-};
-
-export type TerraMetaData = {
-  lastProcessedHeight: number;
-  lastProcessedTimestamp: number;
-  targetHeight: number;
-  chain: string; // Was a bug in the dictionary, should have been chainId
-  indexerHealthy: boolean;
-  indexerNodeVersion: string;
-  queryNodeVersion: string;
 };