+++ packages/utils/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Incorrect type for rowCountEstimate on metadata, remove terra metadata type (#1839)
 
 ## [2.4.1] - 2023-06-09
 ### Changed